Refactor import paths for ProductMedia component to improve module organization
